# docs/v1/openapi.yml
openapi: 3.1.0
info:
  title: Ocha API
  version: 1.0.0
  description: |
    API REST pour la gestion de matcha shops Ocha, incluant la gestion des utilisateurs, produits, magasins et commandes.
    
    ## Fonctionnalités
    - **Gestion de l'authentification** : Inscription et connexion des utilisateurs avec JWT.
    - **Gestion des produits** : Création, mise à jour, suppression et récupération des produits.
    - **Gestion des magasins** : Création, mise à jour, suppression et récupération des magasins.
    - **Gestion des commandes** : Création de commandes par les utilisateurs, mise à jour du statut
    - **Statistiques des commandes** : Récupération des statistiques des commandes par utilisateur.

    ## Authentification
    - **Inscription** : les utilisateurs peuvent s'inscrire en fournissant une adresse e-mail, un mot de passe et un nom d'affichage, eventuellement un numéro de téléphone.
    - **Connexion** : les utilisateurs peuvent se connecter avec leur adresse e-mail et mot de passe pour obtenir un token JWT.
    - **Sécurité** : les endpoints protégés nécessitent un token JWT dans l'en-tête `Authorization Bearer <token>`.
    - **Rôles** : les utilisateurs ont des rôles (`user`, `admin`) qui contrôlent l'accès aux fonctionnalités.
    - **Hashage des mots de passe** : les mots de passe sont stockés de manière sécurisée en utilisant bcrypt.

    ## Code de status de réponse HTTP pouvant être rencontrés
    - `200 OK` : Requête réussie (ressource retournée).
...

...

...

...

...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Store:
      type: object
      required:
        - id
        - name
        - slug
        - email
        - address
        - location
        - is_active
        - opening_hours
        - created_at
      properties:
        id:
          type: string
          format: mongo-id
          description: Identifiant unique.
          example: "64c9f2e4f6a7b8c90d1e2f3"
        name:
          type: string
          minLength: 3
          maxLength: 50
          example: "Ocha Neuchâtel Gare"
        slug:
          type: string
          description: Identifiant lisible dans l'URL, unique.
          example: "ocha-neuchatel-gare"
        phone:
          type: string
          nullable: true
        email:
          type: string
          format: email
        address:
          type: object
          required:
            - line1
            - city
            - zipcode
            - country
          properties:
            line1:
              type: string
            city:
              type: string
            zipcode:
              type: string
            country:
              type: string
          example: {
            line1: "Avenue de la Gare 1",
            city: "Neuchâtel",
            zipcode: "2000",
            country: "Suisse"
          }
        location:
          type: object
          required:
            - type
            - coordinates
          properties:
            type:
              type: string
              enum: [Point]
            coordinates:
              type: array
              minItems: 2
              maxItems: 2
              items:
                type: number
          example: {
            type: "Point",
            coordinates: [6.9896, 46.9910]
          }
        is_active:
          type: boolean
          default: true
        opening_hours:
          type: array
          minItems: 7
          maxItems: 7
          items:
            type: array
            nullable: true
            items:
              type: string
              pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    ####################
    # ORDER
    ####################
...
paths:
...
  /stores:
    get:
      tags:
        - Stores
      summary: Récupère la liste des magasins
      description: Récupère les magasins avec pagination et possibilité de filtrer par proximité géographique.
      operationId: getStores
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: near
          schema:
            type: string
          description: Coordonnées sous forme "longitude,latitude" pour filtrer les magasins proches
        - in: query
          name: radius
          schema:
            type: integer
          description: Rayon de recherche en mètres
      responses:
        "200":
          description: Liste des magasins récupérée
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "List of stores"
                  page:
                    type: integer
                    example: 1
                  totalPages:
                    type: integer
                    example: 1
                  totalStores:
                    type: integer
                    example: 5
                  stores:
                    type: array
                    items:
                      $ref: "#/components/schemas/Store"
              example:
                message: "List of stores"
                page: 1
                totalPages: 1
                totalStores: 5
                stores:
                  - _id: "695fb9529b09cba35cad0044"
                    name: "Ocha Lausanne Flon"
                    slug: "ocha-lausanne-flon"
                    phone: "+41214445566"
                    email: "lausanne-flon@ocha.ch"
                    is_active: true
                    address:
                      line1: "Rue de Genève 21"
                      city: "Lausanne"
                      zipcode: "1003"
                      country: "Suisse"
                    location:
                      type: "Point"
                      coordinates: [6.629, 46.522]
                    opening_hours:
                      - []
                      - ["09:00","18:30"]
                      - ["09:00","18:30"]
                      - ["09:00","18:30"]
                      - ["09:00","20:00"]
                      - ["09:00","20:00"]
                      - ["10:00","18:00"]
                    created_at: "2026-01-08T14:04:02.656Z"
                    updated_at: "2026-01-08T14:04:02.656Z"
                    __v: 0
                  - _id: "695fb9529b09cba35cad0045"
                    name: "Ocha Neuchâtel Gare"
                    slug: "ocha-neuchatel-gare"
                    phone: "+41214445567"
                    email: "neuchatel-gare@ocha.ch"
                    is_active: true
                    address:
                      line1: "Avenue de la Gare 1"
                      city: "Neuchâtel"
                      zipcode: "2000"
                      country: "Suisse"
                    location:
                      type: "Point"
                      coordinates: [6.9896, 46.9910]
                    opening_hours:
                      - []
                      - ["08:30","18:00"]
                      - ["08:30","18:00"]
                      - ["08:30","18:00"]
                      - ["08:30","19:00"]
                      - ["08:30","19:00"]
                      - ["09:00","17:00"]
                    created_at: "2026-01-08T14:10:00.000Z"
                    updated_at: "2026-01-08T14:10:00.000Z"
                    __v: 0
        "400":
          description: Bad request
        "500":
          description: Internal server error

    post:
      tags:
        - Stores
      summary: Crée un nouveau magasin
      description: Crée un magasin Ocha avec toutes ses informations (nom, email, téléphone, adresse, localisation et horaires d'ouverture).
      operationId: createStore
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - address
                - location
                - opening_hours
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                phone:
                  type: string
                  nullable: true
                address:
                  type: object
                  required:
                    - line1
                    - city
                    - zipcode
                    - country
                  properties:
                    line1:
                      type: string
                    city:
                      type: string
                    zipcode:
                      type: string
                    country:
                      type: string
                location:
                  type: object
                  required:
                    - type
                    - coordinates
                  properties:
                    type:
                      type: string
                      enum: [Point]
                    coordinates:
                      type: array
                      minItems: 2
                      maxItems: 2
                      items:
                        type: number
                opening_hours:
                  type: array
                  minItems: 7
                  maxItems: 7
                  items:
                    type: array
                    nullable: true
                    items:
                      type: string
                      pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
            example:
              name: "Ocha Neuchâtel"
              email: "neuchatel@ocha.ch"
              phone: "12345678"
              address:
                line1: "Avenue de la Gare 1"
                city: "Neuchâtel"
                zipcode: "2000"
                country: "Suisse"
              location:
                type: "Point"
                coordinates: [6.93, 46.99]
              opening_hours:
                - []
                - ["08:00", "18:00"]
                - ["08:00", "18:00"]
                - ["08:00", "18:00"]
                - ["08:00", "20:00"]
                - ["08:00", "20:00"]
                - ["09:00", "17:00"]
      responses:
        "201":
          description: Magasin créé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Store created"
                  store:
                    $ref: "#/components/schemas/Store"
              example:
                message: "Store created"
                store:
                  _id: "6962e0fd6aaa44672169d804"
                  name: "Ocha Neuchâtel"
                  slug: "ocha-neuchatel"
                  phone: "12345678"
                  email: "neuchatel@ocha.ch"
                  is_active: true
                  address:
                    line1: "Avenue de la Gare 1"
                    city: "Neuchâtel"
                    zipcode: "2000"
                    country: "Suisse"
                  location:
                    type: "Point"
                    coordinates: [6.93, 46.99]
                  opening_hours:
                    - []
                    - ["08:00","18:00"]
                    - ["08:00","18:00"]
                    - ["08:00","18:00"]
                    - ["08:00","20:00"]
                    - ["08:00","20:00"]
                    - ["09:00","17:00"]
                  created_at: "2026-01-10T23:30:05.756Z"
                  updated_at: "2026-01-10T23:30:05.757Z"
                  __v: 0
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "409":
          description: Conflict
        "422":
          description: Validation error
        "500":
          description: Internal server error
  /stores/{id}:
    get:
      tags:
        - Stores
      summary: Récupère un magasin par ID
      description: Récupère les informations détaillées d'un magasin spécifique.
      operationId: getStoreById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Magasin trouvé
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Store found"
                  store:
                    $ref: "#/components/schemas/Store"
              example:
                message: "Store found"
                store:
                  _id: "695fb9529b09cba35cad0044"
                  name: "Ocha Lausanne Flon"
                  slug: "ocha-lausanne-flon"
                  phone: "+41214445566"
                  email: "lausanne-flon@ocha.ch"
                  is_active: true
                  address:
                    line1: "Rue de Genève 21"
                    city: "Lausanne"
                    zipcode: "1003"
                    country: "Suisse"
                  location:
                    type: "Point"
                    coordinates: [6.629, 46.522]
                  opening_hours:
                    - []
                    - ["09:00","18:30"]
                    - ["09:00","18:30"]
                    - ["09:00","18:30"]
                    - ["09:00","20:00"]
                    - ["09:00","20:00"]
                    - ["10:00","18:00"]
                  created_at: "2026-01-08T14:04:02.656Z"
                  updated_at: "2026-01-08T14:04:02.656Z"
                  __v: 0
        "400":
          description: Bad request
        "404":
          description: Store not found
        "500":
          description: Internal server error

    patch:
      tags:
        - Stores
      summary: Met à jour un magasin
      description: Met à jour les informations d'un magasin existant, comme le nom, l'adresse, la localisation ou les horaires d'ouverture.
      operationId: updateStore
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                phone:
                  type: string
                  nullable: true
                address:
                  type: object
                  required:
                    - line1
                    - city
                    - zipcode
                    - country
                  properties:
                    line1:
                      type: string
                    city:
                      type: string
                    zipcode:
                      type: string
                    country:
                      type: string
                location:
                  type: object
                  required:
                    - type
                    - coordinates
                  properties:
                    type:
                      type: string
                      enum: [Point]
                    coordinates:
                      type: array
                      minItems: 2
                      maxItems: 2
                      items:
                        type: number
                opening_hours:
                  type: array
                  minItems: 7
                  maxItems: 7
                  items:
                    type: array
                    nullable: true
                    items:
                      type: string
                      pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
            example:
              name: "Ocha Neuchâtel Gare"
              address:
                line1: "Avenue de la Gare 2"
                city: "Neuchâtel"
                zipcode: "2000"
                country: "Suisse"
              location:
                type: "Point"
                coordinates: [6.931444286600357, 46.99321666053446]
              opening_hours:
                - []
                - ["08:00","18:00"]
                - ["08:00","18:00"]
                - ["08:00","18:00"]
                - ["08:00","18:00"]
                - ["08:00","20:00"]
                - ["08:00","20:00"]
      responses:
        "200":
          description: Magasin mis à jour avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Store updated"
                  store:
                    $ref: "#/components/schemas/Store"
              example:
                message: "Store updated"
                store:
                  _id: "6962e0fd6aaa44672169d804"
                  name: "Ocha Neuchâtel Gare"
                  slug: "ocha-neuchatel-gare"
                  phone: "12345678"
                  email: "neuchatel@ocha.ch"
                  is_active: true
                  address:
                    line1: "Avenue de la Gare 2"
                    city: "Neuchâtel"
                    zipcode: "2000"
                    country: "Suisse"
                  location:
                    type: "Point"
                    coordinates: [6.931444286600357, 46.99321666053446]
                  opening_hours:
                    - []
                    - ["08:00","18:00"]
                    - ["08:00","18:00"]
                    - ["08:00","18:00"]
                    - ["08:00","18:00"]
                    - ["08:00","20:00"]
                    - ["08:00","20:00"]
                  created_at: "2026-01-10T23:30:05.756Z"
                  updated_at: "2026-01-10T23:32:59.994Z"
                  __v: 1
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
        "409":
          description: Conflict
        "422":
          description: Validation error
        "500":
          description: Internal server error

    delete:
      tags:
        - Stores
      summary: Supprime un magasin
      description: Supprime un magasin existant par son identifiant.
      operationId: deleteStore
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Store deleted successfully
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Store not found
        "500":
          description: Internal server error
  # ---------------------------------------
  # Orders
  # ---------------------------------------
...